refactor(admin): colocate setup wizard and installation flow module
The first-run onboarding lived scattered across view/, composables/, styles/ and setup/. Move the wizard page, each step's logic and scss beside the step components, relocate reusable setup fields/progress dialogs under components/setup, and rewire imports so the whole install flow forms one self-contained module.
