+++ webui/src/admin/components/setup/InstallationSuccessDialog.vue
