+++ webui/src/admin/components/setup/SetupComponentHeader.vue
