+++ webui/src/admin/components/setup/SetupCredentialInput.vue
