+++ webui/src/admin/components/setup/SetupSourceChoice.vue
