+++ webui/src/admin/components/setup/installation-progress.scss
