+++ webui/src/admin/setup/detailed-config-step.scss
