+++ webui/src/admin/setup/mode-selection.scss
