+++ webui/src/admin/setup/setup-complete.scss
