+++ webui/src/admin/setup/setup-wizard.scss
