+++ webui/src/admin/setup/useInstallationProgress.ts
